FIX: moving utils/shared around
